+++ modules/service/src/main/scala/lucuma/odb/graphql/mapping/MutationMapping.scala
@@ -50,7 +50,6 @@ import lucuma.odb.graphql.input.CreateProgramInput
 import lucuma.odb.graphql.input.CreateTargetInput
 import lucuma.odb.graphql.input.CreateUserInvitationInput
 import lucuma.odb.graphql.input.LinkUserInput
-import lucuma.odb.graphql.input.ObservationPropertiesInput
 import lucuma.odb.graphql.input.RecordAtomInput
 import lucuma.odb.graphql.input.RecordDatasetInput
 import lucuma.odb.graphql.input.RecordGmosStepInput
@@ -281,32 +280,10 @@ trait MutationMapping[F[_]] extends Predicates[F] {
     }
 
   private lazy val CreateObservation: MutationField =
-    MutationField("createObservation", CreateObservationInput.Binding) { (input, child) =>
-      services.useTransactionally {
-
-        def createObservation(pid: Program.Id): F[Result[(Observation.Id, Query)]] =
-          observationService.createObservation(pid, input.SET.getOrElse(ObservationPropertiesInput.Create.Default)).map(
-            _.fproduct(id => Unique(Filter(Predicates.observation.id.eql(id), child)))
-          )
-
-        def insertAsterism(pid: Program.Id, oid: Observation.Id): F[Result[Unit]] =
-          input.asterism.toOption.traverse { a =>
-            asterismService.insertAsterism(pid, NonEmptyList.one(oid), a)
-          }.map(_.getOrElse(Result.unit))
-
-        val query = for {
-          pid <- ResultT(programService.resolvePid(input.programId, input.proposalReference, input.programReference))
-          tup <- ResultT(createObservation(pid))
-          (oid, query) = tup
-          _   <- ResultT(insertAsterism(pid, oid))
-        } yield query
-
-        for {
-          rQuery <- query.value
-          _      <- transaction.rollback.unlessA(rQuery.hasValue)
-        } yield rQuery
-      }
-    }
+    MutationField("createObservation", CreateObservationInput.Binding): (input, child) =>
+      services.useTransactionally:
+        observationService.createObservation(input).nestMap: oid =>
+          Unique(Filter(Predicates.observation.id.eql(oid), child))
 
   private lazy val CreateProgram =
     MutationField("createProgram", CreateProgramInput.Binding) { (input, child) =>
